Simplify payment class.
