+++ Components/Blisstribute/Order/SyncMapping.php
@@ -113,7 +113,6 @@ class Shopware_Components_Blisstribute_Order_SyncMapping extends Shopware_Compon
     }
 
     /**
-     * @throws Shopware_Components_Blisstribute_Exception_OrderPaymentMappingException
      * @throws Shopware_Components_Blisstribute_Exception_OrderShipmentMappingException
      * @throws Shopware_Components_Blisstribute_Exception_ValidationMappingException
      * @throws NonUniqueResultException
